test: add generateRandonIds test
