--- src/app/components/superheroes/components/randomSuperheroes/getRandomNumberBetween.ts
@@ -1,2 +1,0 @@
-export const getRandomNumberBetween = (minValue: number, maxValue: number) =>
-  Math.floor(Math.random() * (maxValue - minValue)) + minValue;
